Move RepositoryView to shared
