+++ appshared/build.gradle
@@ -70,4 +70,5 @@ dependencies {
     androidTestCompile 'org.mockito:mockito-core:1.10.19'
     androidTestCompile "com.google.dexmaker:dexmaker:1.2"
     androidTestCompile "com.google.dexmaker:dexmaker-mockito:1.2"
+    compile 'com.github.bumptech.glide:glide:3.7.0'
 }
